requirements.txt
ccxt
pandas
numpy
python-dotenv
plotly
streamlit
websocket-client
openpyxl

